silent install intel MPI

+++ install_hdf5.sh
@@ -1,22 +1,21 @@
-########
-# hdf5 #
-########
+#####################
+# Make Dependencies #
+#####################
 
 # sudo apt install -q -y libhdf5-dev
 
 # szip dependency
 if [[ ! -d szip-2.1.1 ]]; then
-    wget https://support.hdfgroup.org/ftp/lib-external/szip/2.1.1/src/szip-2.1.1.tar.gz
+    if [[ ! -a szip-2.1.1.tar.gz ]]; then
+        wget https://support.hdfgroup.org/ftp/lib-external/szip/2.1.1/src/szip-2.1.1.tar.gz
+    fi
     tar -xvzf szip-2.1.1.tar.gz
     rm -f szip-2.1.1.tar.gz 
 fi
 cd szip-2.1.1
 mkdir /home/ubuntu/szip
 ./configure --prefix=/home/ubuntu/szip
-make >> /dev/null
-if [[! $?]] 
-    
-fi
+make
 make check
 make install
 cd ..
@@ -28,16 +27,21 @@ rm -rf szip-2.1.1
 # tar -xvzf zlib-1.2.11.tar.gz
 
 # hdf5
-wget https://support.hdfgroup.org/ftp/HDF5/current/src/hdf5-1.10.1.tar
-tar -xvf hdf5-1.10.1.tar
-rm -f hdf5-1.10.1.tar
+if [[ ! -d szip-2.1.1 ]]; then
+    if [[ ! -a hdf5-1.10.1.tar ]]; then
+        wget https://support.hdfgroup.org/ftp/HDF5/current/src/hdf5-1.10.1.tar
+    fi
+    tar -xvf hdf5-1.10.1.tar
+    rm -f hdf5-1.10.1.tar
+fi
 
 export CC=icc
 export F9X=ifort
 export CXX=icpc
 
 cd hdf5-1.10.1
-./configure --prefix=/usr/local/hdf5-1.10.1 --enable-fortran --enable-cxx
+mkdir /home/ubuntu/hdf5-1.10.1
+./configure --prefix=/home/ubuntu/hdf5-1.10.1 --enable-fortran --enable-cxx
 make
 make check
 make install
